bot/requirements.txt
discord.py==2.1.1
nostr
python-dotenv
SQLAlchemy
pyyaml
pytz
PyMySQL==1.0.2